+++ tests/testthat/test_adjusted_actuals.R
@@ -3,9 +3,9 @@ context('adjusted.actuals')
 test_that('adjusted.actuals returns correct values',{
 
   data(hook)
-  data<- hook[,1:2]
-  coord<- hook[,3:4]
-  aa<-adjusted.actuals(data,coord,plot.maps=F)
+  data<- hook[ , 1:2]
+  coord<- hook[ , 3:4]
+  aa<-adjusted.actuals(data, coord, plot.maps = FALSE)
 
   expect_match(class(aa),'numeric')
 
@@ -17,15 +17,22 @@ test_that('adjusted.actuals returns correct values',{
 test_that('fails correctly',{
 
   data(hook)
-  data <- hook[,1:2]
-  coord <- cbind(seq(0,50, length.out = dim(hook)[1]),
-                 seq(0,50, length.out = dim(hook)[1]))
+  data <- hook[ , 1:2]
+  coord <- cbind(seq(0, 50, length.out = dim(hook)[1]),
+                 seq(0, 50, length.out = dim(hook)[1]))
 
-  expect_error(adjusted.actuals(data,coord,plot.maps=F),
+  expect_error(adjusted.actuals(data, coord, plot.maps = FALSE),
                'coordinates not integer')
 
-  coord <- cbind(1:200,1:200)
-  expect_error(adjusted.actuals(data,coord,plot.maps=F),
+  coord <- cbind(1:200, 1:200)
+  expect_error(adjusted.actuals(data, coord, plot.maps = FALSE),
                'coordinates and data have different dimensions')
+
+  coord<- hook[ , 3:4]
+
+  expect_warning(adjusted.actuals(data, coord,
+                                  plot.maps = TRUE,
+                                  color.maps = TRUE),
+                 '"color.maps" is now soft deprecated')
 
 })