+++ tests/testthat/test_rvi_plot.R
@@ -8,6 +8,7 @@ for.test <-rvi.plot(carlina.horrida ~ aridity + land.use, "poisson",
                     carlinadata, coords, maxlevel = 4,
                     detail = TRUE, wavelet = "d4")
 
+library(ggplot2)
 
 test_that("outputs are as expected", {
   skip_on_cran()
@@ -16,6 +17,15 @@ test_that("outputs are as expected", {
   expect_equal(as.vector(for.test$rvi[2, ]), c(1, 1, 1, 0),
                tolerance = 1e-5)
   expect_true(is.list(for.test))
+
+  expect_is(for.test$plot, 'ggplot')
+
+  expect_warning(rvi.plot(carlina.horrida ~ aridity + land.use, "poisson",
+                          carlinadata, coords, maxlevel = 4,
+                          detail = TRUE, wavelet = "d4",
+                          customize_plot = ggplot2::theme(legend.position = 'top')),
+                 regexp = '"customize_plot" argument is now soft deprecated')
+
 
 })